job/pyproject.toml
[tool.poetry]
name = "job"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
arxiv = "^1.4.8"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
google-cloud-storage = "^2.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
